feat: 클러스터링 조회 API 구현
- lat, lng 테이블 데이터 타입 DECIMAL -> INT 로 변경

+++ monicar-control-center/src/main/resources/schema/init-schema.sql
@@ -110,8 +110,8 @@ CREATE TABLE cycle_info
     `vehicle_id`    BIGINT       NOT NULL COMMENT '차량 PK',
     `interval_at`   TIMESTAMP    NOT NULL COMMENT '발생시간',
     `status`        VARCHAR(100) NOT NULL COMMENT 'GPS 상태',
-    `lat`           DECIMAL      NOT NULL COMMENT '위도 * 1000000 한값(소수점 6자리)',
-    `lng`           DECIMAL      NOT NULL COMMENT '경도 * 1000000 한값(소수점 6자리)',
+    `lat`           INT          NOT NULL COMMENT '위도 * 1000000 한값(소수점 6자리)',
+    `lng`           INT          NOT NULL COMMENT '경도 * 1000000 한값(소수점 6자리)',
     `ang`           INT          NOT NULL COMMENT '방향 - 범위 : 0 ~ 365',
     `spd`           INT          NOT NULL COMMENT '속도 - 범위 : 0~ 255(단위: km/h)',
     `created_at`    TIMESTAMP    NOT NULL COMMENT '테이블 생성 시간'
